Adding batch script to submit ctest jobs for nightly testing on Hopper.
Also modified CMakeLists.txt so if Albany has been built w/ MPI, the serial executables will be run
with the mpirun or aprun machine (depending on the architecture), like the parallel tests.  This is needed
to automate testing on machines like Hopper, where the executable cannot be run in serial w/o the mpirun/aprun command.