Added a new BC name, "basal_scalar_field", for FELIX problems run with either ASCII/Extruded meshes or through the CISM-Albany and MPAS-Albany interfaces. The BC is for problems where the basal sliding field is a scalar field that lives at all the nodes at the basal boundary.  The specification is as follows:
 <Parameter name="NBC on SS Basal for DOF all set basal_scalar_field" type="Array(double)" value="{1.0}"/>

to set a BC of d(flux)/dx = scale*beta*u, where scale = 1.0 in this case.

This is equivalent to what was previously specified as:

<Parameter name="NBC on SS Basal for DOF all set basal" type="Array(double)" value="{0.0, 0.0, 0.0, 0.0, 0.0}"/>
<Parameter name="BetaXY" type="string" value="Scalar Field"/>

The old specification of the BC still works; but it is recommended that it not be used b/c the input list (Array(double)) is not used in the code.

Also made some clean ups of the basal BC: if an Array(double) of length less than 5 is given in the input file,
the non-specified values are set to 0.0 (to prevent a possible segmentation fault).  If an Array(double) of length > 5 is
given in the input file, an error is thrown.