Some code cleanup.
reduced the allowable mininum jacobian determinant value.

reordered storage of spherical_coordinates to be the commonly accepted
long-lat (lambda, theta) order.

removed Aeras_ShallowWaterSource evaluator in favor of a generalized coriolis
force.  This force reduces to the physical term if alpha = 0.

The test inputs in TC2 and TC5 that used this source were modified
appropriately.

Removed specificationof alpha from input.xml.  This value was hard
coded in the ShallowWaterResidualDef and had to match what was in the
xml.  The issue is that I do not know how to get a parameter into the
AnalyticFunctions that are used by the initial conditions and into the
residual evaluators. By removing it from the input file at least it
cannot be changed by an unsuspecting user, but must be changed in the
code in two places.