Incremental commit toward Peridigm coupling.
Created a PeridigmManager to handle the interface between
Albany and Peridigm.  It is a singleton that is called
directly from Albany::Application for initialization,
updating the current displacements (i.e., copying to the
Peridigm data structures), and computing the peridynamic
internal force.  Nodal forces can they be queried downstream
from an evaluator.  The initial implementation will function
only for Residual evaluations.