More progress on Albany-Peridigm coupling.
Hooked up plumbing to connect the Albany state manager
with the corresponding Peridigm data structures.  If
Peridigm output is requested in the input deck, space
for the data is now allocated on the Albany side and
the values are filled in the PeridigmForce evaluator
(the data is copied from the Peridigm data structures
to the Albany data structures).  The data then appears
in the Exodus output file.

Numerous things are hard-coded and will need to be
generalized.

The PeridigmStaticElasticity3D example problem appears to
run correctly at this point, at least in serial.