Modifying tests for subdivision and recovery
Looking at multiple tests to probe the subdivision code and also investigate tet4 and tet10 elements at the crack tip. Both the fields and the recovered fields are under investigation.

Note to Alejandro: Meshes are now symmetric.

Checking these in now because we will be using these tests to collaborate with Julian Rimoli.