Fixed bug in PeridigmProblem related to reading the input deck.
The Elastic Modulus and Poissons Ratio were not being properly fed
to the ElasticModulus and PoissonsRatio evaluators.  This caused
them to use their default values, which are 1.0 for both Young's modulus
and Poissons ratio.

The tests containing both a peridynamics block and a classical
elasticity block now run correctly.  Prior to the bug fix, they
were failing to converge because the elastic moduli for the
two bars differed by eleven orders of magnitude, plus the Poisson's
ratio was bogus for the FEM bar.