LCM: clean up and add a test for stabilized pressure
This is pretty neat, much more testing to do here.

For anyone interested, take a look in the StabilizedTet4 test directory for
an example for how to turn it on.

Note: this doesn't work yet for the linear elastic material, nor does it work
yet for surface elements. Each of those should be relatively easy extensions.