Beginning to put in logic to build only the AlbanyT executable.  The idea is there will be a flag in the Albany cmake script that can be turned off to turn off building of Albany (and AlbanySG, AlbanyAnalysis, etc. -- all the stuff that needs Epetra).
-D ENABLE_ALBANY_EPETRA_EXE:BOOL=OFF

Ultimately it should be possible to build just AlbanyT w/o having Epetra on in Trilinos.

By default the flag is on (so Albany gets built).  If you try to turn it off now there will be linking
errors, as I have not ifdef'ed all the relevant things out yet.