Fix failing tests in Tpetra branch
Set floor of 1.e-10 on stress in Gurson-shear.exodiff to fix failure due
to relative tolerance on small numbers on some platforms.

Fix ScalarT sqrt() of zero FPE in TvergaardHutchinsonModel_Def.hpp

Fix floating point exception in Intrepid in PHAL_HeatEqResid_Def.hpp.
Intrepid operates on the full worksetSize of data - if the evaluator is
not spanning the entire workset it must still contain reasonable data or
an FPE is thrown in intrepid on some platforms.