Ifdef'ed out Epetra from PHAL_Workset, related functions that still had Epetra, and adapt functions if ALBANY_EPETRA_EXE is turned off.
Some capabilities in adapt are Epetra only -- need to check with Glen which ones need to be
kept (converted to Tpetra).

Converted PHAL_SeparableScatterScalarResponse_Def.hpp .

Now, all the Epetra that should remain if ALBANY_EPETRA_EXE is turned off is some Epetra_Comms which
need to be converted to Teuchos::Comms.  AFter this it should be possible to compile AlbanyT with ALBANY_EPETRA_EXE
turned off and no Epetra in Trilinos (modulo possible bug fixed for some places where I missed Epetra
when I was ifdef-ing/removing it).