Ifdef'ed a few more Epetras I had missed earlier.  Now the command
nm --print-file-name --demangle --undefined-only libalbanyLib.a | grep Epetra | less

yields nothing (and same for the other libraries built as a part of Albany)!  Unfortunately it does not seem possible to build Trilinos w/o Epetra
because various packages that are needed in Albany (e.g., Piro, Zoltan2, Stratimikos, etc.)
are dependent on Epetra.  Need to discuss how to proceed...