Modify the linear and nonlinear convergence tolerances.
Tighten convergance tolerances on both the Newton inner iteration linear solve
and the outer iteration. Now the MeanValue comparison values for -np 1 and -np 4
agree to 11 digits. Intermediate quantities such as Cauchy stress at a given
iteration agree now in a relative normwise sense to 1e-8, down from ~1e-2.

In my previous commit, I had thought that there was a remaining issue in the new
parallel version. The tighter tolerances show that in fact there is not, so I
think I'm now done with parallelization for the moment.