Follow up to previous commit re: response functions using evaluators.
Sacado expects evaluators to follow design (d1). Hence it is appropriate that
Albany::FieldManagerScalarResponseFunction does, too.

However, ProjectIPtoNodalField is a response function that implements EvalT =
PHAL::AlbanyTraits::Residual only. It may be that there is a category of
response functions that follow this pattern. These might include, for example,
transferring data to another module in a loose coupling of Albany with other
software, or writing special files.

I returned Albany::FieldManagerScalarResponseFunction to its previous
implementation and instead created a small subclass of it called Albany::
FieldManagerResidualOnlyResponseFunction. It overloads evaluateTangentT and
evaluateGradientT. These simply call FieldManagerScalarResponseFunction::
evaluateResponseT if gT is not null. Derivatives are 0.

I set ProjectIPtoNodalField and IPtoNodalField to use this new class.