Make AlbanyT and a few other executables build shared.
These mods get AlbanyT and a few other executables to build with shared
libraries. To do this, make Trilinos with
    -D BUILD_SHARED_LIBS:BOOL=ON
There are still some linking errors for a few test executables that I need to
resolve. The important thing is to have the main programs build as shared for
faster rebuilding when doing dev/debug.