Get MOR tests to run with Albany (but not AlbanyT).
Add a warning to Albany/CMakeLists.txt that MOR works only with Albany exe.

Make use of ALBANY_64BIT_INT #define guards more precise in
    Albany::GenericSTKFieldContainer<Interleaved>::fillVectorHelper
Now the various implementations of that routine run fully regardless of the
value of ALB_64BIT_INT.

Add the 3 MOR tests to the nightly tests.