Get Aeras/ShallowWater/TC2 to work when ALBANY_EPTRA_EXE is OFF.
The reason this test was failing is there was a bare abs call.

I haven't traced out the header differences between when ALBANY_EPTRA_EXE is ON
and OFF, but when it's on, double std::abs(double) is called, and when it's not,
int abs(int) was; output of these calls was verified in both cases. I don't
actually fully understand how std::abs is being called in the ON case: first, I
used gdb to confirm it is; and, second, I confirmed that the std namespace isn't
open here. Somehow the compiler is willing to resolve the symbol to std::abs
anyway. Similarly, I haven't figured out why int abs(int) was being called in
the OFF case, but it definitely was. So the exact details will probably have to
remain a mystery.

The solution is to make std::abs explicit.

While I was fixing this problem, I grepped for " abs(" and "(abs(" to see if
there were any other instances of this issue. I found and fixed two.