Fix FPE issues in Laplace-Beltrami test problems.
Main_Solve.cpp used the wrong ifdef guard for the floating point exception
code. Fixed.

TPSLaplaceResid_Def.hpp and similar files were causing FPE because Intrepid
operates on part of the jacobian array that is not actually used, and so
contains invalid values. I followed the exampled in J2Stress to avoid this
problem.