[Part 1 of merging distributed parameters code.] [Builds but not all tests pass.]
Albany: Modified StateInfoStruct and several files in Albany/disc  to allow for distributed nodal fields (or parameters) defined on parts of STK mesh.

The name (string) of the mesh part where the field is defined  is stored in StateInfoStruct.
The enum "NodalDistParameter" has been added to StateStruct::MeshFieldEntity.

A subset of the StateInfoStruct provided by the function GenericSTKFieldContainer::buildStateStructs(const Teuchos::RCP<Albany::StateInfoStruct>& sis) is stored in "nodal_parameter_sis".
The object "nodal_parameter_sis" contains only the StateStrucs of type "NodalDistParameter".

Information contained in "nodal_parameter_sis" is used in STKDiscretization class to (for each parameter):
1. Extract the nodes of the mesh corresponding to the parameter mesh part and create corresponding unique and overlapped Epetra maps.
2. Build a "DofManager", associated to each map that return local ids and global ids.
3. Build workset indices "wsElNodeID" that maps (Elem,node,numComps) into the corresponding local Id of the parameter. The id will be negative when the parameter is not defined on a particular (eleme,node).

These data are stored in class NodalDOFsStructContainer, and, for each parameter, it is possible to access to members of object DOFsStruct using the parameter name.

Same Maps/workset indices and dofManagers can be shared between many parameters to avoid storage duplications.

The solution maps (for the ordinary case), and the node maps are contained in NodalDOFsStructContainer and are built together with other parameters' data.