Get tests corresponding to Error 1 working.
Was just a small copy-paste type bug.

I removed the TensileTet10* tests from the broken list, but those
were fixed in the nodal graph commits.

I added a new set of failing tests (just 1 at the moment) along with a
hypothesis about what the problem is. If I'm right about the cause, then it's in
master, too.