Address compiler warning.
The compiler was emitting: "warning: returning reference to temporary" for,
e.g.,
  shards::CellTopology const&
  get_cell_topology() { return get_topology().get_cell_topology(); }.
In general, one cannot return references to stack variables, though it sometimes
works. I see no reason to keep this code as it was. In particular,
shards::CellTopology is implemented using the Pimpl idiom (aka, d-pointer,
opaque pointer); the data of this object is simply a single pointer. Hence it
can be passed efficiently by value.