Get FO_GIS_Gis*Unstructured tests running.
This fix had multiple parts.

1. Mauro committed a fix to master that made the tests pass there. I
cherry-picked that commit earlier, but the tests still did not pass in merge.

2. Next I fixed a typo in figuring out numComponents in
Albany_ExtrudedSTKMeshStruct that was causing that variable to be uninitialized.

3. Then there was a map issue in Adapt::NodalDataVector. Turns out the global
number of elements was not being reported correctly. I fixed this problem in
Adapt::NodalDataBlock, too. My understanding is that these are new features, so
it's not surprising the test coverage of these is very low.

4. Then I got a seg fault in
    SeparableScatterScalarResponse<PHAL::AlbanyTraits::Jacobian, Traits>::
    evaluateFields.
It turns out that all the evaluator-based response functions that use Albany::
FieldManagerScalarResponseFunction have Tpetra-only impls: in particular, they
read and write T versions of the workset fiedls. However, evaluateGradient (not
T) is (correctly) called by Albany::ModelEvaluator::evalModel. To deal with this
discrepancy, I do Epetra-Tpetra conversions in
    Albany::FieldManagerScalarResponseFunction::evaluateGradient
and call evaluateGradientT. I wrote a comment about this in the code. In the
future, we might strip out the unused code if I'm right that it's no longer
needed.

5. Finally, the sensitivity output was not correct. I fixed a replace/sumInto
bug in
    SeparableScatterScalarResponse<PHAL::AlbanyTraits::Jacobian, Traits>::
    evaluateFields
that fixed this.

I updated the broken-TODO list.