Get SG_MP merged and working.
1. As Irina suggested, I did Tpetra-Epetra conversion and call evaluateResponseT
and evaluateTangentT. I introduced the convenience class Petra::Converter to
simplify this and possibly future conversion tasks. This got Albany to build
with SG_MP on.

2. Comment out over-eager throws. Evaluator constructors need not throw on an
unimplemented EvalT; it is sufficient to throw in evaluateFields. The first will
cause a needless throw if SG_MP is compiled in, say; the latter, in contrast,
will throw only if SG_MP is both compiled in and accidentally used for the given
problem. The second is the desired behavior. I modified two Aeras evaluator
constructors to follow the second behavior; this fixes some Aeras tests that
don't use SG_MP but nonetheless break when SG_MP is turned on.

3. Initialized comm from commT earlier in Albany::Application::Application, as
comm is used in finalSetUp when SG_MP is on.

All previously working tests pass, and now the SG tests pass, too.