[Final part of merging distributed parameters code.]
Mods to make KfieldBC_L2Proj_Tpetra work again. KfieldBC_L2Proj_Tpetra is the
only test that seems to critically test the code path that depends on
Albany::STKDiscretization::getWsElNodeID(). This test showed that
Albany::DOFsStruct needed to be modified to account for wsElNodeEqID and
wsElNodeID separately. This commit does that.