Fix nfm indexing issue.
nfm was being indexed using wsPhysIndex like fm. Current Albany behavior does
not tie nfm indexing to fm's; indeed, it appears nfm is only ever 1, though
perhaps this is not true in uncommitted problems.

Thanks to Jake and James for providing a breaking case, and (earlier) Josh for
addressing this issue in master. In the merge we lost Josh's changes. Until I
understand what everyone wants, I've retained the array and protected the
conflicting behavior in the new function deref_nfm. If nfm.size() == 1, nfm[0]
is returned as desired; if nfm.size() > 1, then nfm[wsPhysIndex[ws]] is
returned.