Strip TimeDebBC.
TimeDepBC is a small modification of the behavior implemented in
PHAL::Dirichlet. Previously, TimeDepBC_Def.hpp copy/pasted all of the code in
PHAL::Dirichlet and then modified it a bit. I've stripped that out and replaced
it with an inheritance structure that uses the PHAL::Dirichlet impl. TimeTracBC
was already doing the right thing.