Modify LCM::TimeDepBC to account for a deforming mesh.
When the mesh deforms as part of reference configuration updating, the
time-dependent Dirichlet BC has to account for the moving boundary.

Right now, the modification is a kludge. A correct PHAL::Workset::previous_time
would be one way to obviate the kludge.

All behavior remains the same unless this new functionality is explicitly
invoked using
  <ParameterList name="Time Dependent DBC on NS ns_4 for DOF Y">
    <Parameter name="Mesh Deforms" type="bool" value="true"/>
  </ParameterList>