Modified lateral BC in PHAL_Neumann class such that the gravitational constant, ice density and water density are taken from the input *.xml file. The syntax in the *.xml file is as follows (under "Problem"):
  <ParameterList name="Physical Parameters">
      <Parameter name="Gravity" type="double" value="9.8"/>
      <Parameter name="Ice Density" type="double" value="910.0"/>
      <Parameter name="Water Density" type="double" value="1028.0"/>
    </ParameterList>

Also changed input array for lateral BC.  The syntax now is as follows:

<Parameter name="NBC on SS lateralside for DOF all set lateral" type="Array(double)" value="{n}"/>

where n is an integer, and should be set to 0 if you want the ImmersedRatio (for an ice sheet calculation)
to be calculated within the code from the thickness and elevation.  If this value is not 0, ImmersedRatio will
be set to a constant value given by n (needed for CISM-Albany problems for example).

The way I modified PHAL_Neumann, the input syntax should be backward compatible with the old syntax that took
an array of 5 doubles.  The last 4 entries in the array will not be used, and probably should not be provided,
to prevent confusion.