Temporary hack to address problem of observer/RF/file-write ordering.
When using LOCA, there are four tightly coupled actions: (a) update the state
field manager, (b) update the solution field, (c) call the response functions in
case they affect the Exodus data to be written, and (d) write data to the Exodus
file.

At present I believe the only correct order of these four in general is (a, b),
c, d; a and b can be interchanged. Albany-LOCA-Epetra currently has the order c,
a, b, d. AlbanyT-LOCA-Thyra currently has the order a, b, d, c. This commit
makes Albany-LOCA-Thyra implement the order a, b, c, d.

There are two main issues.

The first is to distinguish between writing a solution to the mesh database and
writing fields to file. The first must happen approximately in the LOCA
observer. The second has to happen after everything else. For this reason, I
split writeSolutionT into writeSolutionToMeshDatabaseT and
writeSolutionToFileT. (writeSolutionT still does both for use in LCM unit test
executables.)

The second is dealing with the two main ways of accessing the solution at each
step in LOCA: printSolution, which leads to ObserverImpl; and
evalConvergedModel, which leads to ModelEvaluatorT::evalModelImpl. The problem
is that actions a, b, and d happened in ObserverImpl, while action c happens in
evalModelImpl. That means that one could place c before or after a, b, d, but
not between any of them. I think I need to redesign some of this in the
future. I think that will require touching a fair amoung of LOCA and Piro
code. For now I've opted to implement a hacky fix almost entirely (except for
one line of code in Piro_LOCASolver_Def.hpp) inside of Albany. Excerpt from a
comment:

  //              ... If there are response functions and get_g() is not null,
  // then that means we're evaluating response functions after a continuation
  // step or NOX solve. Do the Exodus write here instead of in observeSolutionT.
  //   If Ng() == 0, then this hack isn't needed since there are no response
  // functions.
  //   Main_SolveT determines whether the Exodus file is written here or in
  // observeSolutionT.

Results are nominal for the 324-test test set. Albany and AlbanyT IP-to-nodal
fields differ in the Exodus files for the demo can problem because of the
difference in order between Albany (c, a, b, d) and AlbanyT with the temporary
fix (a, b, c, d). I believe AlbanyT is now correct in this regard and Albany is
not, but I may be wrong.