Hooked up the MultiPhysicsCrystalPlasticity model and added a pair of tests.
The tests are copies of the CrystalPlasticity tests created by Reese Jones, et. al,
with a few tweaks:

- Switched from ifpack to ifpack2 so the tests will run with AlbanyT
- Named the solution "displacement" and the residual "force" for easier ParaView-ing
- Added Regression Results so that there's something to test.
- Hooked things up in the CMakeLists.txt files so the tests will run via the test suite.

Note that the Regression Results currently test the average displacement, which
doesn't make too much sense because the displacements are fully prescribed.  For
future work, the test should be converted to run exodiff.