Make shared build work when library Hydride is enabled.
If Hydride is enabled, then the add_subdirectory statements for LCM and Hydride
have to be handled carefully. LCM's CMakeLists.txt builds several unit test
executables. Hence if Hydride is enabled and shared libs are being built, then
the Hydride library must be available to the LCM unit test link lines in order
to resolve the Hydride-related symbols. (In the case of static libs, the Hydride
symbols are not used, so the static linker is happy to not worry about their
being unresolved.)