Initial commit of reference configuration manager.
Maintain a field x_rc, the reference configuration for the solution x, that is
updated x_rc += x (and then interpolated to the new mesh) at the time of mesh
adapatation.

If this capability is not explicitly turned on in the parameter file, it does
nothing. It is turned on as follows:
  <ParameterList name="Adaptation">
    <Parameter name="Reference Configuration: Update" type="bool" value="true"/>
  </ParameterList>