Force a typecast of 0 to ScalarT.
Bug symptom: Jacobian matrix was accumulating rather than being set to 0 at each
gather.

Cause: The line was
    residual_(cell, node, dim) = 0.0;
This triggered the following call stack:

    at Sacado_Fad_ViewStorage.hpp:104
    operator= (this=0x7fffffff6280, v=@0x7fffffff6278)
    at Sacado_Fad_GeneralFadImp.hpp:98
    (this=0x7fffffff6280, v=@0x7fffffff6278)
    at Sacado_Fad_ViewFad.hpp:150

Problem: Impl of this resize is empty. Hence after
    residual_(cell, node, dim) = 0.0
completes, its state for EvalT=Jacobian is
    0 [a lot of old data]
instead of
    0 [].
I think that's correct since this type isn't allowed to change size. However,
I'm thinking what is needed is something else to implement the semantics of this
operator= overload properly. One possibility: zero'ing the array of derivative
data, so we'd end up with this:
    0 [0 0 ... 0]
Not sure.

In any case, this is in the Kokkos-Sacado code path. On the Albany-Kokkos side,
we can just force the right behavior by constructing a zero of the right type
and thereby invoking a different operator=:
    residual_(cell, node, dim) = typename EvalT::ScalarT(0.0);
A possible problem with this approach is we'll have to be sure that every
instance of
    persistent ScalarT variable = 0.0
is tranformed.

It appears that the can problem is now running correctly.