Adding amb:: debug output routines.
For the next few weeks, it will be useful to have comprehensive debug output to
do correctness checks and bisect where things are going wrong. I've added two
files, amb.hpp and amb.cpp, that implement some usesful routines and
macros. I've also included debug output in some of the evaluators.

Unless I did something wrong, the debug output is set to do nothing right
now. Changing amb::print_level() to 1 for some residual and solution vector
output, 2 for a little more, and >= 11 for Jacobian information.

Output files are all ASCII. I have some Matlab routines I'm not committing (yet)
that read them, but they are pretty trivial formats, so they can be visually
inspected as well.

There are two other things in this commit. amb::set_xT_for_debug(), if it is
changed to return true, tells Albany to inject a fixed xT vector. That is useful
if we want an interesting vector on which to run the evaluators in the Kokkos
and master versions for comparison.

amb::set_own_omp_nthreads, if it is changed to return true, initializes Kokkos
to use 4 OpenMP threads instead of all that is available. No need to fire up 40
threads when debugging.