Make GO long long int rather than long to match Trilinos.
The key benefit of this change is that in our Trilinos configuration, we can now
use
    -D Trilinos_ENABLE_EXPLICIT_INSTANTIATION:BOOL=ON
for both 32-bit (which already worked) and 64-bit int builds.

I updated the instructional Trilinos cmake config file to reflect this change. A
shared + ETI Trilinos build reduces an incremental rebuild of Albany in which
Albany_SolverFactory.cpp (the file that takes perhaps the longest to compile)
must be compiled from ~2 minutes to <40s.

I also reworked some of the bottom portion of the instructional Trilinos cmake
config file to emphasize optimal flag choices when doing dev work.