[Ref config] Framework for RC updating.
Reference configuration computations are done in the AAdapt::rc namespace.

rc::Manager is the interface between the problem, the evaluators, the state
manager, and AAdapt::MeshAdapt. It maintains a database (implemented using
Albany::StateManager) of accumulated quantites (q_accum) that are combined with
incremental quantities (q_incr).

rc::Manager deploys the evaluators rc::Reader and rc::Writer to read from the
database and write back to it. These two evaluators make q_accum quantities
available to the physics evaluators. rc::Reader is used in all field
managers. rc::Writer is used in the state field manager to do the write of the
current step's q_accum.

rc::Field is a wrapper to an MDField that holds the q_accum quantity for a given
q_incr. A physics evaluator containing a q_incr field uses an rc::Field to gain
access to the associated q_accum. rc::Field implements the various operations
needed to combine q_accum and q_incr.

Data types in the AAdapt::rc namespace are defined in
AAdapt_RC_DataTypes.hpp. AAdapt_RC_DataTypes_impl.hpp provides macros to help
with RC-specific ETI; it is meant to be included only in .cpp files.

The BC field managers are given x (the solution, which is displacment) + x_accum
(the accumulated solution). This lets time-dependent BCs work naturally.

Other field managers are given just x.

Quite a bit still has to be done in terms of math, but I want to push this
initial commit now that the framework is established and
tested. ElasticityProblem with the Strain evaluator are the only problem and
q_accum-q_incr pair currently hooked up to AAdapt::rc. The first interesting
problem I'll work on is in fact MechanicsProblem and not ElasticityProblem, but,
as usual, ElasticityProblem provides a simpler setting for dev and testing.