[Ref config] Implement Lie group <-> algebra transformations. Not enabled yet.
1. Infrastructure: rc::Transformation::Enum, rc::Manager::tellAdapted.

2. In rc::Manager, modify reading and writing of fields to permit associating
multiple intermediate fields with one external field. For example, in F = R S, R/r
and S/s must be stored, but F is used externally.

3. Implement rc::Manager::FieldDatabase::transform. Right now, the following
transformations, using Intrepid::MiniTensor, are implemented:

  a. Strain components are directly interpolated.

  b. For deformation gradient F,
    G -> g:
       F -> R S  // right polar decomp
       r <- log_rotation(R)
       s <- log_sym(S)
    g -> G:
       R <- exp_skew_symmetric(r)
       S <- exp(s)
       F <- R*S.

However, I have not yet enabled the transformations in rc::Manager. I have to do
some work, probably in LOCA_AdaptiveStepper.C, to signal a mesh adaptation is
about to occur, and thus q_accum should be transformed G -> g and written to the
field database. Right now an adaptation is assumed to occur at the end of each
continuation step. However, the initial solve does not have an adaptation, and
this catastrophically messes up the transformation code. Hence this commit adds
a bit of code but does not alter the behavior of the previous [Ref config]
commit.

Another chunk of work I still have to do is the L^2 nodal projection. For now
I'm using the SCOREC QP interpolation capability.