Update getIntrepidBasis() function for spectral elements
After Irina's change to check the first few characters of the
CellTopologyData name (which has the effect of ignoring the number of
nodes suffix), I refined the logic to convert the number of nodes to
an element degree.  If this succeeds, I can then call the appropriate
Intrepid::Basis_HGRAD_*_Cn_FEM constructor, which in effect allows for
spectral elements.