Creating a version of QCAD_Solver that uses the Thyra::ModelEvaluator instead of the EpetraExt::ModelEvaluator, towards having QCAD run with the AlbanyT executable.
Switched EpetraExt classes to use Thyra, rewriting relevant functions (e.g., replacing get_g_map with
get_g_space, etc.).  Code compiles but a lot of it still needs to be rewritten with Thyra,
after which it needs to be tested.  There is a lot of use of EpetraExt::ModelEvaluator
functions so the task of converting everything to Thyra::ModelEvaluator is proving to be nontrivial.
It should be doable with some effort, starting with this initial commit.