Added a new response that prints the solution for an array of Node GIDs
This required a new Culling Strategy for the SolutionValues response.
The Node GIDs in the input file are 1-based, as seen in paraview,
so we subtract 1 in the code to get 0-based GIDs.

SolutionValues response are NOT working with Tpetra.

Added a test within SteadyHeat3D.

Pair programmed with John Mitchell.