Creating infrastructure for spectral elements for the Aeras atmospheric model.  The plan for now is to read in a linear quad/hex mesh from Exodus and enrich it within the Aeras::SpectralDiscretization class to be used with Albany.  For output, we plan to use VTK, which supports higher order elements (implementation not started yet specifically for Aeras, but which exists for FMDB: in the class AlbPUMI::FMDBVtk). Right now the Aeras::SpectralDiscretization class is simply a copy of Albany::STKDiscretization.
To call the Aeras::SpectralDiscretization class, the "Method" line in an input file
should read "Ioss Aeras" or "Exodus Aeras" (which are equivalent).  It works now
and does the same thing as Albany::STKDiscretization.

I have begun putting some commenting on what needs to be done in Aeras::SpectralDiscretization
for the enriched elements, but that's not complete yet.  I will add more commenting tomorrow.
I will also remove LCM calls as the SpectralDiscretization will for now be Aeras-specific
and the enrichment limited to certain kinds of elements/meshes (spherical quad/hex).