Fix Tpetra bug in SolutionValuesResponseFunction.
I removed numResponsesT, as it was used only in
SolutionValuesResponseFunction. numResponses is in fact what is called. The
result was that in AlbanyT, the number of responses was reported always as 0,
which was the bug. I rewrote numResponses to take care of both Epetra and Tpetra
cases.

While I was at it, I also removes a bit of duplicated code in
SolutionCullingStrategy that is independent of T/Epetra.

Finally, I added a Tpetra test to SteadyHeat3D that mimics the Epetra one that
Andy added a couple of days ago.