Cleanup Clang test on "avatar", adjust options.
Ran into strange gfortran issue with the "algebra" utility.

"algebra" will not process command line arguments when Seacas and
dependencies are compiled with Clang 3.5/gfortran 4.8.2. Attempted to track this
down. The issue is that the fortran iargc() function always returns
-1 instead of the number of command line arguments. This appears to
be an optimization issue - dropping back to using "-Os" instead of
"-O3" for optimization in Clang and gfortran for the dashboard tests.