LCM: update to the volume averaging procedures for J and pressure
Recent developments and scrutiny of the volume averaging methods have
uncovered issues with accuracy and consistency.

Previously for the determinant of the deformation gradient, J, we took the
volume average of log(J), but that has been shown to produce inconsistent
and non-intuitive behavior. Simply computing the volume average of J gives
us more expected results. That correction has been made to Kinematics_Def.hpp.

Also, We are taking the volume average of the pressure extracted from the
Cauchy stress, a measure that lives in the current configuration. Previously
we were computing the volume average of the pressure with the volume in
the reference configuration, but it was determined that it is more consistent
to use the volume in the current configuration. Fixing this required passing
in J along with the integration weights as optional fields in the
ConstitutiveModel Base class to use in the volumeAverage method.

Thanks to Jay Foulk and Alejandro Mota for sorting this out.