LCM: fix two tests in Mechanics for recent volume average changes
This test (which is duplicated for Albany and AlbanyT), is not ideal in that
there is no analytic solution to compare to. In fact it was really added
just as a functionality test for multiple element blocks with multiple
material models defined. In this case the volume averaging for the J2 block
is set to true, so I need to update the refence solution for the changes.

Note: anyone who is super ambitious could do some work to make this problem
most useful by doing some verification work....