Protect use of long long int in Albany_Memory.cpp by proper ifdef.
I forgot this in my initial commit of this file. I'm using long long int here,
if the Teuchos MPI specializations for it are available, because this routine
tracks memory use, and these numbers can exceed 2^32 regardless of GO (provided
the memory-tracking routine uses 64-bit ints).

Thanks to Alejandro for pointing out this problem.